+++ exportplan/core/serializers.py
@@ -9,13 +9,6 @@ from directory_constants import choices
 from exportplan.utils import format_two_dp
 
 
-class ExportPlanRecommendedCountriesSerializer(serializers.Serializer):
-    sectors = serializers.ListField(child=serializers.CharField())
-
-    def validate_sectors(self, value):
-        return value[0].split(',')
-
-
 class CountryTargetAgeDataSerializer(serializers.Serializer):
     target_age_groups = serializers.ListField(
         required=False, allow_empty=True, allow_null=True, default=[], child=serializers.CharField()
@@ -297,7 +290,6 @@ class TravelBusinessPoliciesSerializer(serializers.Serializer):
 class ExportPlanSerializer(serializers.Serializer):
     export_commodity_codes = ExportPlanCommodityCodeSerializer(many=True, required=False)
     export_countries = ExportPlanCountrySerializer(many=True, required=False)
-    target_markets = serializers.ListField(child=serializers.CharField(), required=False)
     about_your_business = AboutYourBuinessSerializer(required=False)
     objectives = ObjectiveSerializer(required=False)
     target_markets_research = TargetMarketsResearchSerializer(required=False)