+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - GP2-1617: Port get-finance/UKEF contact form from great-domestic-ui
 - GP2-2856  remove unused api calls
 - no-ticket - package upgrade
+- GP2-2856 - remove target market
 
 ### Fixed bugs
 - GP2-2915 - DAC_Information_and_Relationships_06 Roles on WTE tables